autoclickVision/requirements.txt
opencv-python>=4.8
mss>=9.0
pyautogui>=0.9.54
pydirectinput>=1.0.4
PyQt6>=6.6
numpy>=1.24
pyyaml>=6.0
keyboard>=0.13
Pillow>=10.0
requests>=2.31
schedule>=1.2
pyinstaller>=6.0
